+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## to 0.8.x
 
 * Add Argon2 format - Fixes [#84](https://github.com/duritong/trocla/issues/84)
+* Add yescrypt format - Fixes [#81](https://github.com/duritong/trocla/issues/81)
 
 ## to 0.7.0